requirements.txt
alembic==1.13.3
altair==5.4.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.0
attrs==23.2.0
blinker==1.8.2
boto3==1.35.39
botocore==1.35.39
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
cohere==5.11.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.34.0
Deprecated==1.2.14
dill==0.3.9
docker==7.1.0
exceptiongroup==1.2.2
fastavro==1.9.7
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.2
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain-core==0.3.10
langchain-text-splitters==0.3.0
langsmith==0.1.134
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
mkl==2021.4.0
mlflow==2.17.0
mlflow-skinny==2.17.0
mock==4.0.3
mpmath==1.3.0
multiprocess==0.70.17
narwhals==1.9.3
networkx==3.2.1
numpy==1.26.4
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
pathos==0.3.3
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
protobuf==4.25.5
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.11
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==307
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
sagemaker==2.232.2
sagemaker-core==1.0.10
sagemaker-mlflow==0.1.0
schema==0.7.7
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.2.0
six==1.16.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
streamlit==1.39.0
sympy==1.13.3
tbb==2021.13.1
tblib==3.0.0
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.20.1
toml==0.10.2
torch==2.4.1
torchgen==0.0.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
waitress==3.0.0
watchdog==5.0.3
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
